重さ完了 分解　未完成
