+++ Assets/Mituki/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab98c8661091ad849a5fb53dd8cce681
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
