+++ Assets/Mituki/Script/ThrowingScript.cs
