+++ Assets/Mituki/Script/ThrowingScript.cs.meta
