+++ Assets/Mituki/Script/Throwingpower.cs.meta
