+++ Assets/Mituki/Script/kinds.cs.meta
