+++ Assets/Mituki/bottle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ebd536b9016d10d40a5df98296b6c25c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
