+++ Assets/Mituki/cap.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3667cf3674d65d34bac9dfe35a481f19
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
