+++ Assets/Scenes/GameScene.unity
@@ -690,7 +690,7 @@ Transform:
   m_GameObject: {fileID: 878756081}
   serializedVersion: 2
   m_LocalRotation: {x: 0.7071068, y: 0, z: 0, w: 0.7071068}
-  m_LocalPosition: {x: 0, y: -2.1, z: -4.4}
+  m_LocalPosition: {x: 0, y: -4.3, z: -4.4}
   m_LocalScale: {x: 50, y: 80, z: 0.1}
   m_ConstrainProportionsScale: 0
   m_Children: []
