+++ ProjectSettings/TagManager.asset
@@ -15,9 +15,9 @@ TagManager:
   - 
   - Water
   - UI
-  - 
-  - 
-  - 
+  - Plastic_bottle
+  - bottle
+  - cap
   - 
   - 
   -
